pyproject.toml
[project]
name = "gameoflife"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "nifle3", email = "kuper116rus.2005@gmail.com"},
]
dependencies = [
    "pygame-ce>=2.5.1",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
